refactor(events): extract revenue aggregation engine to its own module
